Mais reestruturamentos.

+++ src/main/java/fatec/ads/fabbico/repos/RepoUsuario.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RepoUsuario extends JpaRepository<Usuario, Integer> {
+public interface RepoUsuario extends JpaRepository<Usuario, Long> {
 
     Usuario findUsuarioByNome(String nome);
